# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "Login": {
        "additionalProperties": false,
        "properties": {
          "email": {
            "maxLength": 100,
            "pattern": "^[^@\\s]+@[^@\\s]+\\.[^@\\s]+$",
            "title": "Email",
            "type": "string"
          },
          "password": {
            "format": "password",
            "maxLength": 128,
            "minLength": 12,
            "title": "Password",
            "type": "string",
            "writeOnly": true
          }
        },
        "required": [
          "email",
          "password"
        ],
        "title": "Login",
        "type": "object"
      },
...
      "PageMeta": {
        "properties": {
          "next_cursor": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "required": [
          "next_cursor"
        ],
        "title": "PageMeta",
        "type": "object"
      },
...
      "ProfileUpdate": {
        "additionalProperties": false,
        "properties": {
          "capacity_kg": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*(?:\\d{0,6}|(?=[\\d.]{1,9}0*$)\\d{0,6}\\.\\d{0,2}0*$)",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Capacity Kg"
          },
          "latitude": {
            "anyOf": [
              {
                "maximum": 90.0,
                "minimum": -90.0,
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d{0,6}0*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latitude"
          },
          "longitude": {
            "anyOf": [
              {
                "maximum": 180.0,
                "minimum": -180.0,
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d{0,6}0*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Longitude"
          },
          "name": {
            "anyOf": [
              {
                "maxLength": 100,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "phone": {
            "anyOf": [
              {
                "pattern": "^\\+[1-9][0-9]{7,14}$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phone"
          }
        },
        "title": "ProfileUpdate",
        "type": "object"
      },
...
      "Registration": {
        "additionalProperties": false,
        "properties": {
          "capacity_kg": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*(?:\\d{0,6}|(?=[\\d.]{1,9}0*$)\\d{0,6}\\.\\d{0,2}0*$)",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Capacity Kg"
          },
          "email": {
            "format": "email",
            "maxLength": 100,
            "title": "Email",
            "type": "string"
          },
          "latitude": {
            "anyOf": [
              {
                "maximum": 90.0,
                "minimum": -90.0,
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d{0,6}0*$",
                "type": "string"
              }
            ],
            "title": "Latitude"
          },
          "longitude": {
            "anyOf": [
              {
                "maximum": 180.0,
                "minimum": -180.0,
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d{0,6}0*$",
                "type": "string"
              }
            ],
            "title": "Longitude"
          },
          "name": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "password": {
            "format": "password",
            "maxLength": 128,
            "minLength": 12,
            "title": "Password",
            "type": "string",
            "writeOnly": true
          },
          "phone": {
            "pattern": "^\\+[1-9][0-9]{7,14}$",
            "title": "Phone",
            "type": "string"
          },
          "roles": {
            "items": {
              "enum": [
                "Donor",
                "Receiver",
                "Volunteer"
              ],
              "type": "string"
            },
            "maxItems": 3,
            "minItems": 1,
            "title": "Roles",
            "type": "array"
          },
          "zone_id": {
            "exclusiveMinimum": 0.0,
            "title": "Zone Id",
            "type": "integer"
          }
        },
        "required": [
          "email",
          "password",
          "name",
          "phone",
          "zone_id",
          "roles",
          "latitude",
          "longitude"
        ],
        "title": "Registration",
        "type": "object"
      },
      "RoleData": {
        "properties": {
          "approved": {
            "title": "Approved",
            "type": "boolean"
          },
          "role": {
            "enum": [
              "Donor",
              "Receiver",
              "Volunteer",
              "Admin"
            ],
            "title": "Role",
            "type": "string"
          }
        },
        "required": [
          "role",
          "approved"
        ],
        "title": "RoleData",
        "type": "object"
      },
...
      "SessionData": {
        "properties": {
          "csrf_token": {
            "title": "Csrf Token",
            "type": "string"
          },
          "user": {
            "$ref": "#/components/schemas/UserData"
          }
        },
        "required": [
          "user",
          "csrf_token"
        ],
        "title": "SessionData",
        "type": "object"
      },
      "SessionResponse": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/SessionData"
          }
        },
        "required": [
          "data"
        ],
        "title": "SessionResponse",
        "type": "object"
      },
...
      "UserData": {
        "properties": {
          "active": {
            "title": "Active",
            "type": "boolean"
          },
          "capabilities": {
            "items": {
              "enum": [
                "Donor",
                "Receiver",
                "Volunteer",
                "Admin"
              ],
              "type": "string"
            },
            "title": "Capabilities",
            "type": "array"
          },
          "capacity_kg": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Capacity Kg"
          },
          "email": {
            "title": "Email",
            "type": "string"
          },
          "latitude": {
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Latitude",
            "type": "string"
          },
          "longitude": {
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Longitude",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "phone": {
            "title": "Phone",
            "type": "string"
          },
          "roles": {
            "items": {
              "$ref": "#/components/schemas/RoleData"
            },
            "title": "Roles",
            "type": "array"
          },
          "user_id": {
            "title": "User Id",
            "type": "integer"
          },
          "verified_status": {
            "title": "Verified Status",
            "type": "boolean"
          },
          "zone_id": {
            "title": "Zone Id",
            "type": "integer"
          }
        },
        "required": [
          "user_id",
          "zone_id",
          "name",
          "email",
          "phone",
          "latitude",
          "longitude",
          "verified_status",
          "active",
          "roles",
          "capabilities",
          "capacity_kg"
        ],
        "title": "UserData",
        "type": "object"
      },
      "UserResponse": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/UserData"
          }
        },
        "required": [
          "data"
        ],
        "title": "UserResponse",
        "type": "object"
      },
      "UsersResponse": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/UserData"
            },
            "title": "Data",
            "type": "array"
          },
          "meta": {
            "$ref": "#/components/schemas/PageMeta"
          }
        },
        "required": [
          "data",
          "meta"
        ],
        "title": "UsersResponse",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "Verification": {
        "additionalProperties": false,
        "properties": {
          "reason": {
            "maxLength": 300,
            "minLength": 3,
            "title": "Reason",
            "type": "string"
          },
          "roles": {
            "items": {
              "enum": [
                "Donor",
                "Receiver",
                "Volunteer"
              ],
              "type": "string"
            },
            "maxItems": 3,
            "title": "Roles",
            "type": "array"
          },
          "verified": {
            "title": "Verified",
            "type": "boolean"
          }
        },
        "required": [
          "roles",
          "verified",
          "reason"
        ],
        "title": "Verification",
        "type": "object"
      },
      "ZoneData": {
        "properties": {
          "city": {
            "title": "City",
            "type": "string"
          },
          "zone_id": {
            "title": "Zone Id",
            "type": "integer"
          },
          "zone_name": {
            "title": "Zone Name",
            "type": "string"
          }
        },
        "required": [
          "zone_id",
          "zone_name",
          "city"
        ],
        "title": "ZoneData",
        "type": "object"
      },
      "ZonesResponse": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/ZoneData"
            },
            "title": "Data",
            "type": "array"
          },
          "meta": {
            "$ref": "#/components/schemas/PageMeta"
          }
        },
        "required": [
          "data",
          "meta"
        ],
        "title": "ZonesResponse",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Zone-based food redistribution, accounts, delivery and impact reporting.",
    "title": "Second Table API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/admin/users": {
      "get": {
        "operationId": "admin_users_api_v1_admin_users_get",
        "parameters": [
          {
            "in": "query",
            "name": "verified",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Verified"
            }
          },
          {
            "in": "query",
            "name": "zone_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "exclusiveMinimum": 0,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Zone Id"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Cursor",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UsersResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Admin Users",
        "tags": [
          "Identity",
          "Administration"
        ]
      }
    },
...
    "/api/v1/admin/users/{user_id}/verify": {
      "post": {
        "operationId": "verify_user_api_v1_admin_users__user_id__verify_post",
        "parameters": [
          {
            "in": "path",
            "name": "user_id",
            "required": true,
            "schema": {
              "title": "User Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Verification"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Verify User",
        "tags": [
          "Identity",
          "Administration"
        ]
      }
    },
    "/api/v1/auth/login": {
      "post": {
        "operationId": "login_api_v1_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Login"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Login",
        "tags": [
          "Identity"
        ]
      }
    },
...
    "/api/v1/auth/me": {
      "get": {
        "operationId": "me_api_v1_auth_me_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Me",
        "tags": [
          "Identity"
        ]
      },
      "patch": {
        "operationId": "update_profile_api_v1_auth_me_patch",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProfileUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Profile",
        "tags": [
          "Identity"
        ]
      }
    },
    "/api/v1/auth/register": {
      "post": {
        "operationId": "register_api_v1_auth_register_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Registration"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Register",
        "tags": [
          "Identity"
        ]
      }
    },
...
    "/api/v1/zones": {
      "get": {
        "operationId": "zones_api_v1_zones_get",
        "parameters": [
          {
            "in": "query",
            "name": "city",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "City"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Cursor",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ZonesResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Zones",
        "tags": [
          "Identity"
        ]
      }
    }
  }
}